Add route53 private dns entry for k8s ingress IP
GKE ingress only provide an IP. There's no hostname attached to it like
in aws alb.
This is a problem since test.sh works over https, hence it needs an endpoint
with a tls certificate that match the request Host header and CN.

So, we created a self-signed certificate for *.wso2testgrid.com CN.
This cert's private key is added to the ingrex-nginx. It's public certificate
is added to test.sh client's client-truststore.jks.

Finally, we create a random loadbalancer hostname under .wso2testgrid.com hostname
, and create a dns entry with that for the ingress IP at a route53 private hosted zone.

for ex: for ingress IP 34.67.150.119, we may add a dns entry with hostname wso2am-138.gke.wso2testgrid.com.

        {
            ResourceRecords: [
                {
                    Value: 34.67.150.119
                }
            ],
            Name: wso2am-138.gke.wso2testgrid.com.,
            Type: A,
            TTL: 60
        }